rajout test unitaire

+++ src/main/java/com/mediscreendiabete/model/Note.java
@@ -6,7 +6,7 @@ public class Note {
     private String id;
 
     /** The patient id. */
-    private Long patientId;
+    private int patientId;
 
     /** The note. */
     private String note;
@@ -14,7 +14,7 @@ public class Note {
 	public Note() {
 	}
 
-	public Note(String id, Long patientId, String note) {
+	public Note(String id, int patientId, String note) {
 		super();
 		this.id = id;
 		this.patientId = patientId;
@@ -29,11 +29,11 @@ public class Note {
 		this.id = id;
 	}
 
-	public Long getPatientId() {
+	public int getPatientId() {
 		return patientId;
 	}
 
-	public void setPatientId(Long patientId) {
+	public void setPatientId(int patientId) {
 		this.patientId = patientId;
 	}
